+++ apps/explorer/priv/repo/migrations/20221117075456_modify_address_token_balances_indexes.exs
@@ -13,7 +13,7 @@ defmodule Explorer.Repo.Migrations.ModifyAddressTokenBalancesIndexes do
     create_if_not_exists(
       unique_index(
         :address_token_balances,
-        [:address_hash, :token_contract_address_hash, "COALESCE(token_id, 0)", :block_number],
+        [:address_hash, :token_contract_address_hash, "COALESCE(token_id, -1)", :block_number],
         name: :fetched_token_balances
       )
     )
@@ -29,7 +29,7 @@ defmodule Explorer.Repo.Migrations.ModifyAddressTokenBalancesIndexes do
     create_if_not_exists(
       unique_index(
         :address_token_balances,
-        [:address_hash, :token_contract_address_hash, "COALESCE(token_id, 0)", :block_number],
+        [:address_hash, :token_contract_address_hash, "COALESCE(token_id, -1)", :block_number],
         name: :unfetched_token_balances,
         where: "value_fetched_at IS NULL"
       )